scripts/requirements.txt
boto3==1.4.0
botocore==1.4.53
Fiona==1.6.3.post1
PyYAML==3.11
Rtree==0.8.2
Shapely==1.5.14
Unidecode==0.4.19
click==6.4
click-plugins==1.0.3
cligj==0.4.0
cssselect==0.9.1
distribute==0.7.3
future==0.15.2
futures==3.0.5
ftfy==4.2.0
gevent==1.1.2
greenlet==0.4.10
jmespath==0.9.0
leveldb==0.193
lxml==3.6.0
lru-dict==1.1.3
marisa-trie==0.7.2
numpy==1.10.4
pycountry==1.20
git+https://github.com/kmike/pymorphy2
pymorphy2-dicts-ru==2.4.394633.4298366
pymorphy2-dicts-uk==2.4.1.1.1460299261
pyproj==1.9.5.1
pystache==0.5.4
python-Levenshtein==0.12.0
python-geohash==0.8.5
requests==2.9.1
s3transfer==0.1.3
six==1.10.0
ujson==1.35
urlnorm==1.1.3
wsgiref==0.1.2
